refactor(PR): integrating viewModel
Replaced the mocked travel containers by firestore calls
